+++ connectors/src/types/shared/redis_client.ts
